+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
-## v0.1.0 - v0.1.3 - 20240314
+## v0.1.4 - 20250318
+- 增加 mcp-tavily-api 服务，包括 search、extract 方法
+- 更新 wrangler 版本，改为 pnpm 做包管理；移除 mcp-image-generation 服务的 readMe 方法
+
+## v0.1.0 - v0.1.3 - 20250314
 - 增加 mcp-image-generation 服务
 - 增加 mcp-jina-api 服务，包括 reader、search